+++ source/Nuke.GlobalTool.Tests/UpdateSolutionFileContentTests.cs
@@ -3,8 +3,11 @@
 // https://github.com/GreemDev/NUKE/blob/master/LICENSE
 
 using System;
+using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
+using System.Xml;
+using System.Xml.Linq;
 using Nuke.Common.Utilities;
 using VerifyXunit;
 using Xunit;
@@ -139,5 +142,39 @@ public class UpdateSolutionFileContentTests
 
         return Verifier.Verify(expected)
             .UseParameters(number);
+    }
+    
+    
+    [Theory]
+    [InlineData(
+        1,
+        """
+        <Solution>
+          <Project Path="TestProject1/TestProject1.csproj" />
+        </Solution>
+        """,
+        """
+        <Solution>
+          <Project Path="TestProject1/TestProject1.csproj" />
+          <Project Path="RELATIVE">
+            <Build Project="false" />
+          </Project>
+        </Solution>
+        """)]
+    public Task TestXml(int number, string input, string expected)
+    {
+        var content = XDocument.Load(new StringReader(input));
+        Program.UpdateSolutionXmlFileContent(content, "RELATIVE");
+
+        var settings = new XmlWriterSettings { OmitXmlDeclaration = true, Indent = true };
+        var stringStream = new StringWriter();
+        using var writer = XmlWriter.Create(stringStream, settings);
+        content.Save(writer);
+        writer.Flush();
+
+        if (stringStream.ToString() != expected)
+            Assert.Fail("Unexpected modification to .slnx");
+
+        return Task.CompletedTask;
     }
 }